Move fetch capacity functions to capacities utils file
